Create Eth_Bank.sol
Eth_Bank is a smart contract where anyone can deposit Eth there to be spent later. Users can allow other people to spend for their accounts as well and the contract has an owner and guardians that can change the current owner if necessary (at least 3 guardians need to vote for the new owner).